Updated against PHP 5.5.14
Bunch of new stuff in 5.5.14, I also had to disable the varnish extension as I
couldn't get it working locally. Starting to think that perhaps I should build
out a Linux VM to run the updates against as doing so on OS X has been a bit of
a pain with a handful of packages.